BREAK: mel list, lastmicro_age_days can be None
Change the value of the attribute 'lastmicro_age_days' so that it is
'None' if there are no micro images, instead of an empty string. This
means that the number of columns in stdout is constant.

Extract the functionality to determine lastmicro_age_days to
mel.micro.fs, so that it can be re-used later.