mel.rotomap.relate: new pair-based relate method
Introduce a new autopaste method, based on pairs of moles, that has
similar performance to the 'offset field'. This might cope better with
rotation and scaling between images, as it is based on the vector
between two known moles, rather than the average translation from one
image to another.

Although the results don't appear to be significantly different, it
seems with keeping to develop on in the future.

Test Plan:
.. Update mel-debug bench-autopaste to use the new method ..
.. See that it performs pretty much the same as offset-field ..