Add 'mel rotomap-compare' to compare rotomap moles
In addition to comparing mole images taken with a microscope, it can be
useful to compare mole images taken in a rotomap. If the zoom and
resolution of the camera is high enough then it should be possible to
spot significant differences.

Where moles are inconvenient to image with a microscope, this may be a
good fallback for more regular imaging.

Test Plan:
.. Compare images across a set of rotomaps ..