BREAK: mel rotomap-automask: only arg is TARGET
Vastly improve skin area marking by adopting a simpler method, which
does not require any training. Remove the parameters relating to
training and make the only arguments positional instead.

Take advantage of the structure of the rotomap images, which are mostly
of the skin colour and of the background colour. Since they are expected
to be bi-modal in this way, use Otsu's Method to partition the pixels
into foreground and background, or in other words, skin and non-skin.

This will not work where images are not arranged in this way,
particularly if there are large areas that are of a similar colour to
the skin. In the case of the author, beige walls posed a problem.