Add 'mel rotomap-identify'
Automatically identify moles by walking through the possible
combinations.

Introduce a dependency on scipy, for the kde implementation.

Assign a probability of each mole identity by looking at the nearest
neighbour. Determine probabilities by looking at historical positions of
neighbours and their distribution. Multiply the probability of each
chosen identity together to calculate the cost of each combination. The
combination with the lowest cost is chosen.

As the space of possible combinations is very large, use the 'branch and
bound' approach to reach a solution efficiently. Employ lower-bound
estimates of the future cost of branches to avoid exploring unlikely
combinations.

Use Kernel Density Estimates to determine the probable relative
locations of a mole's neighbours. From each mole, ask it's nearest
neighbour what identities are most likely.

Normalise positions into a skin-mask-relative space. This means that
values are somewhat applicable across images with different resolutions
and camera distance to the area being mapped.

This method doesn't work very well in the presence of new moles, as
there are no likely combinations.

Further work needs to be done to bridge across groups of moles, at the
moment it is possible for small groups of moles to validate each-other
independently of the rest of the image. With less constraints to satisfy
in this way, incorrect combinations are often reached. This is
worked-around by contriving to supply canonical moles in each group.