BREAK: mel rotomap-identify: --target now optional
Move 'TARGET' to be an optional parameter '--target' instead.

This makes the CLI slightly easier to use by allowing re-ordering of
arguments passed.

I find that usually it's the target that needs to be varied - having it
at the end of a previous command line makes it easier to ctrl-w and
supply a new target.