rotomap-edit: mask-attached image transition
If zoomed in, try to preserve the mask-relative position when
transitioning between images.

This makes it much easier to compare mole positions across images when
not in 'follow mode'. This is especially useful when dealing with new
images, which don't yet have any mole UUIDs in common with reference
images.