lowerbound.cpp: calc guesses up-front
If we're going to have any Python exceptions, better to have them all
up-front during initialization time of the CalcGuessesFunctor.

This also removes the need to check if things are in the cache.

Later, it may be useful that the CalcGuessesFunctor is immutable after
construction - we could share it between threads.