mel rotomap-calc-space: better algorithm
It appears that fitEllipse() was a poor choice for this job, as quite a
lot of the resulting ellipses have been way too large, too small, or
off-centre.

Use minAreaRect() instead, this seems to have given much better results.
It seems that some edge-cases remain.

In later work we'll want to choose a different name that 'ellipse' for
this data. Maybe something like 'OBB' for oriented-bounding-box.