Add 'mel rotomap-rm' for removing moles
In the case where it's decided that something wasn't a mole after all,
or is already captured in a different 'part' then it can be desirable to
remove it from the data. This prevents it from needing to be marked in
all the subsequent rotomaps.

Provide a utility for removing moles conveniently.