setup.py: pin pycodestyle for flake8 3.5.0
Somehow we're not resolving the dependency for flake8 correctly
automatically.

Both autopep8 and flake8 depend on pycodestyle.

Currently autopep8 wants >= 2.3, flake8 wants <2.4.0,>=2.0.0.

Pin pycodestyle to the 2.3 series, which will keep both happy for now.

The next release of flake8 should be compatible with >=2.4, so all
should be good again then, and we can unpin.