rotomap-identify: add new mole possiblity
Generate new mole uuids, and add this to the list of possibilities for
each location, except if it's already canonical.

This should make guesses more robust in the presence of spurious or new
moles, which has typically been a problem for generating guesses.

This does highlight the problem of using a new mole as a predictor for
another location. Note that this is also currently not handled well.