rotomap-identify, optimise: only try possibles
Instead of considering the lowerbound of every permutation of identities
and positions, instead only consider those that have more than the
minimum score emitted from calc_guesses.

This shrinks the search space considerably, it seems without reducing
the number of possible outcomes - we can only output permutations of the
emissions of calc_guesses.