rotomap-calc-space: normalize ellipses
Ensure more consistent results by rotation bounding boxes by 90 degrees
(which we're still referring to as 'ellipses'), until the rotation is as
close to zero as we can get.

If the ellipses of the images in a set of rotomaps are not close, then
the results of identification will not be very good.