Add command 'mel rotomap-filter-marks'
Use deep-learning basics taught in the first lesson of fast.ai to
distinguish between things we're tracking as moles and things we aren't.

  http://course.fast.ai/lessons/lesson1.html

Aim to reduce manual labour, this command can be run after
'mel rotomap-automark' to eliminate some false +ves.

Note that in order to use this, you must have created your own model to
run, which is a significant undertaking. In later work we should make it
easier to do this directly from your mel data.

Using 'mel-debug bench-automark', we can see this method can outperform
basic 'mel rotomap-automark' in every way:

    With 'human' vs. 'rotomap-automark + rotomap-filter-marks':
    8234 matches
    2344 falsely marked as moles (false +ve)
    2738 false marked as non-moles (false -ve)

Note that the following rotomap-automark values are with the original
setting, which had a much lower false +ve rate. We changed the values so
that we can get a lower false -ve rate.

    With 'human' vs. 'rotomap-automark':
    7666 matches
    7530 falsely marked as moles (false +ve)
    3306 falsely marked as non-moles (false -ve)

Since the false negatives are largely due to the initial
rotomap-automark, we're not going to be able to improve upon that much
with rotomap-filter-marks. In later work we'll probably want to take a
full-image deep-learning approach to also locate the moles, e.g. YOLO
(You only look once). Handily this is covered in fast.ai deep-learning
part 2.