BREAK: rotomap-edit: swap modes 3 & 4
From a workflow point of view, we do 'bounding area' first, and then
'mole marking'.

Re-arrange the mode buttons to reflect this order.

Also raise an exception if we're somehow in an unknown mode.

In later work it may also make sense to move mole-edit mode to the end,
since we do that last in the workflow. The downside would be that the
new default mode, 'mask edit', would be less useful for browsing.