Add 'mel-debug gen-repo' for developing and tests
Easily generate a usable mel repository, which doesn't have any actual
photographs in it.

This is useful for developing on mel when you don't have any data of
your own, or don't want to expose the somewhat personal, sensitive, and
potentially inappropriate data to where you're working.

It is also useful for generating standard test scenarios, such that they
could be used in continuous integration.