BREAK: rotomap-edit: rm automask in mask edit mode
Not really using this, except by hitting it by accident and messing up
the mask.

Could have instead updated it to use the newer automask routine, but it
seems preferable to always do it in batch mode - without an undo
function it's just too destructive.

Remove it to reduce complexity. There will be lower level code we can
remove in the future, as we gradually remove these high-level references
to it.