rotomap edit: more useful 'unconfirmed' rendering
Improve on the rendering of unconfirmed or non-canonical moles by
making them look the same as confirmed ones, except with some additions.

This makes it easier to visually compare between rotomaps; confirming
that the colours, and therefore uuids, are the same as expected.

The previous, strikingly different, rendering meant that you also had to
account for the difference between concentric circles and concentric
squares.