tricolour: danger colors when exhausted, not crash
Instead of crashing when we run out of colours to use, return danger
colours instead. This means that we can continue with some functionality
in the event of an unreasonable amount of colours being needed.

This has happened in cases where we have a lot of 'auto-marks' that we
want to review.