mel rotomap filter-marks: rebase on fast.ai v1
Now that the 3rd edition of the fast.ai course is live, we can take
advantage of the latest advances, and rebase on the newly released
fastai library.

    https://course.fast.ai/

This allows us to simplify the code somewhat, and we should be able to
benefit from upstream improvements without too much work on our part.