mel rotomap filter-marks*: re-make it, trainable
Re-implement 'mel rotomap filter-marks' such that it's easy to train it
with mel commands, as opposed to having to take a (rather good) deep
learning course and implement a solution for yourself.

Implement it such that a GPU is not necessary, the training can complete
in good time on a laptop with just CPU.

Also, we no longer need the fast.ai library, this is simple enough to do
directly on top of the lower-level PyTorch library.