mel rotomap edit: support looks_like attr editing
For training a classifier, it's useful to be able to annotate the things
we track as things that do or don't look like moles.

For example, sometimes a mole is not really visible, and yet we know
it's location by looking at the previous image and subtle clues in the
current image. It doesn't seem right to confuse the classifier with
these examples if it doesn't have the full context.

Similarly, sometimes we are not tracking some moles, and yet we
certainly want the classifier to treat them as moles.