Add 'mel rotomap identify2*' commands
Introduce an alternative method for identifying moles, using Deep
Learning. This approach seems to yield better results, faster. There is
an additional cost for the initial training though, so training is now
an additional extra command that must be run.

The two new commands are:

    mel rotomap identify2
    mel rotomap identify2-train

Their functionality is somewhat experimental at this point, so they will
keep the 'identify2' moniker until reliable enough to replace the
current solver-based approach.

That said, the current approach does not seem to be stable enough to
exercise using 'system_tests.py', but the new identify2* approach is.
Perhaps there isn't far to go until identify2* can be the only solution.