mel-debug genrepo: raytrace images instead
Use simple raytracing to create more realistic images, paving the way to
better smoke tests.

We'll be able to easily simluate the rotation of a limb with this
method. This means we'll be able to produce multiple images of the same
simulated limb from different angles. We'll then be able test that e.g.
'mel rotomap identify' works as expected, since we can generate fake
'historical data' to train on.