identify2-train: support fine-tuning
Allow 'continuing' training of a previously saved classifier.

Also support re-initialising the fully-connected part, and the
embeddings, if the number of parts or classes differ.

This means that we can pre-train on a large amount of simulated data, to
get better results when training on a smaller amount of real data.