Import PyTorch as late as possible, for speed
It seems that 'import pytorch' introduces a noticable delay. Avoid all
commands incurring this delay by importing only when torch is actually
used.

The multiple 'import torch' statements are maybe a bit objectionable, in
later work we might want to add a layer of indirection, so that we don't
have that peppered multiple times in the same module.