mel status: slightly faster without '-t'
If the user hasn't selected an importance level that will show
InfoNotification items, then don't do some of the work to figure out
what they would be.

This reduces it from 2.2s to 1.2s on my dataset, which has around 4.5k
json files. It seems to be the handling of these json files that takes
the bulk of the time.