rotomap organise: fullscreen with pygame
Start using the `pygame` library to give us a fullscreen display and
associated input events. Fullscreen is something that we couldn't get
with opencv, and I seem to remember seeing a deprecation notice for the
ui part when last installing it on Mac.

Add `pygame` to `setup.py`.

Write a new `mel.lib.fullscreenui` to wrap the bits of `pygame` we need.
This should reduce repetition and help us avoid importing `pygame` in
commands that don't need it. Note that `pygame` prints a banner when
it's imported.

Port `mel.cmd.rotomaporganise` to use the new `fullscreenui`.

Ignore `pygame` in `.pylintrc`, unfortunately PyLint give spurious
errors for this too.

Remove `WAITKEY_BACKSPACE`, now unused.