identifynn: handle duplicate uuids
In the case where we're interactively iterating on the identity of moles
with 'rotomap edit', it seems to happen quite often that 'identify' will
have created some moles with duplicate uuids, and then not update them
again.

Handle the duplicate uuid case by explicitly listing them out, instead
of first converting to Dict[uuid, pos] and then making a list.