identifynn: support non-canonical background moles
If a mole is not marked as canonical, we can use it in training by
having it present when training on the canonical moles, but not training
recognising the non-canonical mole itself.

This means that in combination with the "extra_stem" data, we should be
able to get rotomap-identify to learn to work with spurious moles from
rotomap-automark and friends.