requirements.txt: only specify things in setup.py
It seems that running `pip freeze` on Mac does not always result in a
valid configuration on Linux and v.v.

Try to be a bit more pragmatic, and get less noise from dependabot, by
only freezing the things we explicitly depend on. This does mean that
transitive dependencies may have vulnerabilities that we're not
detecting, but we're also more likely to have time to process the
smaller volume of dependabot updates that we do get. The resulting
`requirements.txt` also might even be valid.