rotomap compare-extra-stem --compare-uuids
Add a `--compare-uuids` argument to `mel rotomap compare-extra-stem`,
for end-to-end performance comparison vs. the 'ground truth'.

This means that you can see how the automatic marking and identification
performs from just an image through to matching the moles in the image
with the marked-up versions.

For example, you can run it like this:

    $ find rotomaps/parts/{Left,Right}*/{Upper,Lower}/2021_* -iname '*.jpg' | xargs mel r compare-extra-stem --error-distance 10 auto --compare-uuids
    matched: 1861
    missing: 1017
    added: 613
    Recall: 65%
    Precision: 75%

Note that the numbers above are from testing on things it was trained
on, so it's not entirely indicative of expected performance on a new
dataset.