micro add --last-changed: use __last_changed__
Make it faster to update moles that have recently changed; re-combine
the capture and compare step as with moles that are not expected to have
changed recently.

By allowing the specification of a time when the mole last changed, we
can compare against that one to see if there have been ongoing changes,
at the same time as capturing the new image.

This seems to speed things up greatly when capturing regular moles.