ZoomableMixin: preserve scale across image size
If the previous image is a different size than the current one, assume
that it is representing the same scene but with a different resolution.

Therefore, try to preserve the apparent size of things on screen by
adjusting the zoom level accordingly.

Reset this scaling whenever switching back the fitted view. This means
that we always get the 'true' resolution when zooming initially.