WIP: rotomapguessrefine: patches as semantic for whole image
Treat the patches as something akin to identifiers for part of the image, with
the understanding that the whole passed image is taken as context for producing
the features for each patch.

In this way it makes more sense to provide lots of context for the SRC patch,
which is simply the central 14x14 patch from a 910x910 context image.

Try to then find the best matching 14x14 patch in the vicinity of the guessed
location in TGT, again using the 910x910 context image (65x65 patches).